slow ci needs more time
we're moving the bbr-cli pipeline to another internally hosted CI.

That means that all of the requests actually have to cross datacenter
boundaries. The instance that runs the bbr cli within these tests is expected
to receive a `killall bbr` before this slow backup job finishes. Unfortunately
it is reliably unable to do so because we see
```
  stdout | bbr: no process found
```
and the tests fail on these two lines:

https://github.com/cloudfoundry/bosh-backup-and-restore/blob/844d7841fb390262b88e9b43cd1d471a2543c526/system/deployment/backup_cleanup_test.go#L34
https://github.com/cloudfoundry/bosh-backup-and-restore/blob/844d7841fb390262b88e9b43cd1d471a2543c526/system/deployment/restore_cleanup_test.go#L53

Basically this tries to make the slow-backup even slower.